# Lokata.WebApi/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Lokata.WebApi, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Document": {
      "get": {
        "tags": [
          "Document"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Document"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Document"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Document"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Document"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Document"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Document"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Document"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Document/{id}": {
      "get": {
        "tags": [
          "Document"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Document"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Document"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Document"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Document"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Document"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Document"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Document"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Document"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Place": {
      "get": {
        "tags": [
          "Place"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Place"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Place"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Place"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Place"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Place"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Place"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Place"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Place/{id}": {
      "get": {
        "tags": [
          "Place"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Place"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Place"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Place"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Place"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Place"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Place"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Place"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Place"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/ScoreCard": {
      "get": {
        "tags": [
          "ScoreCard"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ScoreCard"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ScoreCard"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ScoreCard"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ScoreCard"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScoreCard"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ScoreCard"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ScoreCard"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/ScoreCard/{id}": {
      "get": {
        "tags": [
          "ScoreCard"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ScoreCard"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScoreCard"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScoreCard"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "ScoreCard"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScoreCard"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ScoreCard"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ScoreCard"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "ScoreCard"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Sex": {
      "get": {
        "tags": [
          "Sex"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Sex"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Sex"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Sex"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Sex"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Sex"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Sex"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Sex"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Sex/{id}": {
      "get": {
        "tags": [
          "Sex"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Sex"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Sex"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Sex"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Sex"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Sex"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Sex"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Sex"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Sex"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Address": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "createdBy": {
            "type": "string",
            "nullable": true
          },
          "createdOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "modifiedBy": {
            "type": "string",
            "nullable": true
          },
          "modifiedOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "isDeleted": {
            "type": "boolean"
          },
          "fullName": {
            "type": "string",
            "nullable": true
          },
          "street": {
            "type": "string",
            "nullable": true
          },
          "number": {
            "type": "string",
            "nullable": true
          },
          "postalCode": {
            "type": "string",
            "nullable": true
          },
          "postalPlace": {
            "type": "string",
            "nullable": true
          },
          "places": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Place"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Approach": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "createdBy": {
            "type": "string",
            "nullable": true
          },
          "createdOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "modifiedBy": {
            "type": "string",
            "nullable": true
          },
          "modifiedOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "isDeleted": {
            "type": "boolean"
          },
          "competitionId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "competitionsId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "instructorId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "umpireId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "competition": {
            "$ref": "#/components/schemas/Competition"
          },
          "competitions": {
            "$ref": "#/components/schemas/Competitions"
          },
          "instructor": {
            "$ref": "#/components/schemas/Instructor"
          },
          "scoreCards": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScoreCard"
            },
            "nullable": true
          },
          "takePlaces": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TakePlace"
            },
            "nullable": true
          },
          "targetsAndCardsPhotos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TargetsAndCardsPhoto"
            },
            "nullable": true
          },
          "targetsOrCardTakes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TargetsOrCardTake"
            },
            "nullable": true
          },
          "umpire": {
            "$ref": "#/components/schemas/Umpire"
          }
        },
        "additionalProperties": false
      },
...
      "Competitions": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "createdBy": {
            "type": "string",
            "nullable": true
          },
          "createdOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "modifiedBy": {
            "type": "string",
            "nullable": true
          },
          "modifiedOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "isDeleted": {
            "type": "boolean"
          },
          "placeId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "date": {
            "$ref": "#/components/schemas/DateOnly"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "approaches": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Approach"
            },
            "nullable": true
          },
          "place": {
            "$ref": "#/components/schemas/Place"
          }
        },
        "additionalProperties": false
      },
      "Competitor": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "createdBy": {
            "type": "string",
            "nullable": true
          },
          "createdOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "modifiedBy": {
            "type": "string",
            "nullable": true
          },
          "modifiedOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "isDeleted": {
            "type": "boolean"
          },
          "dateOfBirth": {
            "$ref": "#/components/schemas/DateOnly"
          },
          "age": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "isDisabledPerson": {
            "type": "boolean",
            "nullable": true
          },
          "professional": {
            "type": "boolean",
            "nullable": true
          },
          "sexId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "fullName": {
            "type": "string",
            "nullable": true
          },
          "scoreCards": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScoreCard"
            },
            "nullable": true
          },
          "sex": {
            "$ref": "#/components/schemas/Sex"
          },
          "takePlaces": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TakePlace"
            },
            "nullable": true
          },
          "targetsAndCardsPhotos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TargetsAndCardsPhoto"
            },
            "nullable": true
          },
          "targetsOrCardTakes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TargetsOrCardTake"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DateOnly": {
        "type": "object",
        "properties": {
          "year": {
            "type": "integer",
            "format": "int32"
          },
          "month": {
            "type": "integer",
            "format": "int32"
          },
          "day": {
            "type": "integer",
            "format": "int32"
          },
          "dayOfWeek": {
            "$ref": "#/components/schemas/DayOfWeek"
          },
          "dayOfYear": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "dayNumber": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
...
      "Document": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "createdBy": {
            "type": "string",
            "nullable": true
          },
          "createdOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "modifiedBy": {
            "type": "string",
            "nullable": true
          },
          "modifiedOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "isDeleted": {
            "type": "boolean"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "filename": {
            "type": "string",
            "nullable": true
          },
          "fileContent": {
            "type": "string",
            "format": "byte",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Place": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "createdBy": {
            "type": "string",
            "nullable": true
          },
          "createdOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "modifiedBy": {
            "type": "string",
            "nullable": true
          },
          "modifiedOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "isDeleted": {
            "type": "boolean"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "shootingPlacesCount": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "addressNavigation": {
            "$ref": "#/components/schemas/Address"
          },
          "competitionsList": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Competitions"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ScoreCard": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "createdBy": {
            "type": "string",
            "nullable": true
          },
          "createdOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "modifiedBy": {
            "type": "string",
            "nullable": true
          },
          "modifiedOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "isDeleted": {
            "type": "boolean"
          },
          "competitorId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "approachId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "score": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "approach": {
            "$ref": "#/components/schemas/Approach"
          },
          "competitor": {
            "$ref": "#/components/schemas/Competitor"
          }
        },
        "additionalProperties": false
      },
      "Sex": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "createdBy": {
            "type": "string",
            "nullable": true
          },
          "createdOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "modifiedBy": {
            "type": "string",
            "nullable": true
          },
          "modifiedOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "isDeleted": {
            "type": "boolean"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "asMale": {
            "type": "boolean"
          },
          "asFemale": {
            "type": "boolean"
          },
          "competitorList": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Competitor"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}